Refactor pin entry dialog
Move code for handling input into a specific child fragment which handles text mode
This will help prepare for a second mode of input
