+++ padlock/src/main/java/com/pyamsoft/padlock/pin/PinEntryDialog.java
@@ -18,35 +18,23 @@ package com.pyamsoft.padlock.pin;
 
 import android.annotation.SuppressLint;
 import android.app.Dialog;
-import android.content.Context;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.DialogFragment;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.Window;
 import android.view.WindowManager;
-import android.view.inputmethod.EditorInfo;
-import android.view.inputmethod.InputMethodManager;
-import android.widget.EditText;
 import android.widget.Toast;
 import com.pyamsoft.padlock.Injector;
 import com.pyamsoft.padlock.PadLock;
 import com.pyamsoft.padlock.R;
 import com.pyamsoft.padlock.databinding.DialogPinEntryBinding;
 import com.pyamsoft.padlock.iconloader.AppIconLoaderPresenter;
-import com.pyamsoft.padlock.list.LockListFragment;
 import com.pyamsoft.padlock.lock.common.LockTypePresenter;
-import com.pyamsoft.padlock.model.event.PinEntryEvent;
-import com.pyamsoft.padlock.lock.master.MasterPinSubmitCallback;
-import com.pyamsoft.pydroid.ActionSingle;
 import com.pyamsoft.pydroid.drawable.AsyncDrawable;
 import com.pyamsoft.pydroid.drawable.AsyncMap;
 import com.pyamsoft.pydroid.drawable.AsyncMapEntry;
@@ -55,62 +43,15 @@ import timber.log.Timber;
 
 public class PinEntryDialog extends DialogFragment {
 
+  @NonNull public static final String TAG = "PinEntryDialog";
+  @NonNull static final String CHILD_TAG = "Child_PinEntry";
   @NonNull private static final String ENTRY_PACKAGE_NAME = "entry_packagename";
   @NonNull private static final String ENTRY_ACTIVITY_NAME = "entry_activityname";
-  @NonNull private static final String CODE_DISPLAY = "CODE_DISPLAY";
-  @NonNull private static final String CODE_REENTRY_DISPLAY = "CODE_REENTRY_DISPLAY";
-  @NonNull private static final String HINT_DISPLAY = "HINT_DISPLAY";
   @NonNull private final AsyncMap taskMap = new AsyncMap();
-  @SuppressWarnings("WeakerAccess") InputMethodManager imm;
-  @SuppressWarnings("WeakerAccess") @Inject PinEntryPresenter presenter;
+  @SuppressWarnings("WeakerAccess") @Inject LockTypePresenter presenter;
   @SuppressWarnings("WeakerAccess") @Inject AppIconLoaderPresenter appIconLoaderPresenter;
   DialogPinEntryBinding binding;
-  EditText pinEntryText;
-  EditText pinHintText;
   private String packageName;
-  private EditText pinReentryText;
-
-  @NonNull final PinEntryPresenter.SubmitCallback submitCallback =
-      new PinEntryPresenter.SubmitCallback() {
-
-        @Override public void onSubmitSuccess() {
-          clearDisplay();
-          dismiss();
-        }
-
-        @Override public void onSubmitFailure() {
-          clearDisplay();
-          dismiss();
-        }
-
-        @Override public void onSubmitError() {
-          clearDisplay();
-          dismiss();
-        }
-
-        @Override public void handOffPinEvent(@NonNull PinEntryEvent event) {
-          actOnLockList(callback -> {
-            switch (event.type()) {
-              case 0:
-                if (event.complete()) {
-                  callback.onCreateMasterPinSuccess();
-                } else {
-                  callback.onCreateMasterPinFailure();
-                }
-                break;
-              case 1:
-                if (event.complete()) {
-                  callback.onClearMasterPinSuccess();
-                } else {
-                  callback.onClearMasterPinFailure();
-                }
-                break;
-              default:
-                throw new RuntimeException("Invalid event type: " + event.type());
-            }
-          });
-        }
-      };
 
   public static PinEntryDialog newInstance(final @NonNull String packageName,
       final @NonNull String activityName) {
@@ -164,41 +105,16 @@ public class PinEntryDialog extends DialogFragment {
 
     presenter.initializeLockScreenType(new LockTypePresenter.LockScreenTypeCallback() {
       @Override public void onTypeText() {
-
+        // Push text as child fragment
+        getChildFragmentManager().beginTransaction()
+            .replace(R.id.pin_entry_dialog_container, new PinEntryTextFragment(), CHILD_TAG)
+            .commitNow();
       }
 
       @Override public void onTypePattern() {
-
+        // TODO
       }
     });
-
-    // Resolve TextInputLayout edit texts
-    pinEntryText = binding.pinEntryCode.getEditText();
-    if (pinEntryText == null) {
-      throw new NullPointerException("No pin entry edit text");
-    }
-
-    pinReentryText = binding.pinReentryCode.getEditText();
-    if (pinReentryText == null) {
-      throw new NullPointerException("No pin re-entry edit text");
-    }
-
-    pinHintText = binding.pinHint.getEditText();
-    if (pinHintText == null) {
-      throw new NullPointerException("No pin hint edit text");
-    }
-
-    // Force the keyboard
-    imm = (InputMethodManager) getContext().getApplicationContext()
-        .getSystemService(Context.INPUT_METHOD_SERVICE);
-    imm.toggleSoftInput(InputMethodManager.SHOW_FORCED, InputMethodManager.HIDE_IMPLICIT_ONLY);
-
-    clearDisplay();
-    setupGoArrow(submitCallback);
-
-    if (savedInstanceState != null) {
-      onRestoreInstanceState(savedInstanceState);
-    }
   }
 
   private void setupCloseButton() {
@@ -211,88 +127,6 @@ public class PinEntryDialog extends DialogFragment {
         .tint(android.R.color.black)
         .into(binding.pinEntryClose);
     taskMap.put("close", task);
-  }
-
-  void setupSubmissionView(@NonNull EditText view,
-      @NonNull PinEntryPresenter.SubmitCallback submitCallback) {
-    view.setOnEditorActionListener((textView, actionId, keyEvent) -> {
-      if (keyEvent == null) {
-        Timber.e("KeyEvent was not caused by keypress");
-        return false;
-      }
-
-      if (keyEvent.getAction() == KeyEvent.ACTION_DOWN && actionId == EditorInfo.IME_NULL) {
-        Timber.d("KeyEvent is Enter pressed");
-        presenter.submit(getCurrentAttempt(), getCurrentReentry(), getCurrentHint(),
-            submitCallback);
-        return true;
-      }
-
-      Timber.d("Do not handle key event");
-      return false;
-    });
-  }
-
-  private void setupGoArrow(@NonNull PinEntryPresenter.SubmitCallback submitCallback) {
-    binding.pinImageGo.setOnClickListener(view -> {
-      presenter.submit(getCurrentAttempt(), getCurrentReentry(), getCurrentHint(), submitCallback);
-      imm.toggleSoftInputFromWindow(getActivity().getWindow().getDecorView().getWindowToken(), 0,
-          0);
-    });
-
-    // Force keyboard focus
-    pinEntryText.requestFocus();
-
-    final AsyncMapEntry task =
-        AsyncDrawable.load(R.drawable.ic_arrow_forward_24dp).into(binding.pinImageGo);
-    taskMap.put("arrow", task);
-  }
-
-  private void onRestoreInstanceState(@NonNull Bundle savedInstanceState) {
-    Timber.d("onRestoreInstanceState");
-    final String attempt = savedInstanceState.getString(CODE_DISPLAY, null);
-    final String reentry = savedInstanceState.getString(CODE_REENTRY_DISPLAY, null);
-    final String hint = savedInstanceState.getString(HINT_DISPLAY, null);
-    if (attempt == null || reentry == null || hint == null) {
-      Timber.d("Empty attempt");
-      clearDisplay();
-    } else {
-      Timber.d("Set attempt %s", attempt);
-      pinEntryText.setText(attempt);
-      Timber.d("Set reentry %s", reentry);
-      pinReentryText.setText(reentry);
-      Timber.d("Set hint %s", hint);
-      pinHintText.setText(hint);
-    }
-  }
-
-  @Override public void onSaveInstanceState(@NonNull Bundle outState) {
-    Timber.d("onSaveInstanceState");
-    outState.putString(CODE_DISPLAY, getCurrentAttempt());
-    outState.putString(CODE_REENTRY_DISPLAY, getCurrentReentry());
-    outState.putString(HINT_DISPLAY, getCurrentHint());
-    super.onSaveInstanceState(outState);
-  }
-
-  /**
-   * Clear the display of all text entry fields
-   */
-  void clearDisplay() {
-    pinEntryText.setText("");
-    pinReentryText.setText("");
-    pinHintText.setText("");
-  }
-
-  @SuppressWarnings("WeakerAccess") @CheckResult @NonNull String getCurrentAttempt() {
-    return pinEntryText.getText().toString();
-  }
-
-  @SuppressWarnings("WeakerAccess") @CheckResult @NonNull String getCurrentReentry() {
-    return pinReentryText.getText().toString();
-  }
-
-  @SuppressWarnings("WeakerAccess") @CheckResult @NonNull String getCurrentHint() {
-    return pinHintText.getText().toString();
   }
 
   @Override public void onStart() {
@@ -310,21 +144,6 @@ public class PinEntryDialog extends DialogFragment {
             dismiss();
           }
         });
-    presenter.hideUnimportantViews(new PinEntryPresenter.HideViewsCallback() {
-      @Override public void showExtraPinEntryViews() {
-        Timber.d("No active master, show extra views");
-        binding.pinReentryCode.setVisibility(View.VISIBLE);
-        binding.pinHint.setVisibility(View.VISIBLE);
-        setupSubmissionView(pinHintText, submitCallback);
-      }
-
-      @Override public void hideExtraPinEntryViews() {
-        Timber.d("Active master, hide extra views");
-        binding.pinReentryCode.setVisibility(View.GONE);
-        binding.pinHint.setVisibility(View.GONE);
-        setupSubmissionView(pinEntryText, submitCallback);
-      }
-    });
   }
 
   @Override public void onStop() {
@@ -343,22 +162,11 @@ public class PinEntryDialog extends DialogFragment {
 
     Timber.d("Destroy AlertDialog");
     taskMap.clear();
-    imm.toggleSoftInputFromWindow(getActivity().getWindow().getDecorView().getWindowToken(), 0, 0);
     binding.unbind();
   }
 
   @Override public void onDestroy() {
     super.onDestroy();
     PadLock.getRefWatcher(this).watch(this);
-  }
-
-  void actOnLockList(@NonNull ActionSingle<MasterPinSubmitCallback> action) {
-    final FragmentManager fragmentManager = getFragmentManager();
-    final Fragment lockListFragment = fragmentManager.findFragmentByTag(LockListFragment.TAG);
-    if (lockListFragment instanceof LockListFragment) {
-      ((LockListFragment) lockListFragment).provideMasterSubmitCallback(action);
-    } else {
-      throw new ClassCastException("Fragment is not MasterPinSubmitCallback");
-    }
   }
 }
