+++ padlock/src/main/res/layout/dialog_pin_entry.xml
@@ -67,102 +67,11 @@
           android:contentDescription="@string/blank"
           />
 
-      <RelativeLayout
+      <FrameLayout
+          android:id="@+id/pin_entry_dialog_container"
           android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          >
-
-        <ImageView
-            android:id="@+id/pin_image_go"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:layout_alignParentEnd="true"
-            android:layout_centerVertical="true"
-            android:layout_marginStart="8dp"
-            android:contentDescription="@string/blank"
-            />
-
-        <ScrollView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentStart="true"
-            android:layout_toStartOf="@id/pin_image_go"
-            >
-
-          <LinearLayout
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:orientation="vertical"
-              >
-
-            <android.support.design.widget.TextInputLayout
-                android:id="@+id/pin_entry_code"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:gravity="center_horizontal"
-                android:hint="@string/lock_code_hint"
-                >
-
-              <com.pyamsoft.pydroid.design.widget.NoLeakTextInputEditText
-                  android:id="@+id/pin_entry_code_text"
-                  android:layout_width="match_parent"
-                  android:layout_height="match_parent"
-                  android:gravity="center_horizontal"
-                  android:imeOptions="actionNext"
-                  android:inputType="textPassword"
-                  android:textAppearance="?android:attr/textAppearanceMedium"
-                  android:textColor="@android:color/black"
-                  />
-
-            </android.support.design.widget.TextInputLayout>
-
-            <android.support.design.widget.TextInputLayout
-                android:id="@+id/pin_reentry_code"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:gravity="center_horizontal"
-                android:hint="@string/reenter_lock_code_hint"
-                >
-
-              <com.pyamsoft.pydroid.design.widget.NoLeakTextInputEditText
-                  android:id="@+id/pin_reentry_code_text"
-                  android:layout_width="match_parent"
-                  android:layout_height="match_parent"
-                  android:gravity="center_horizontal"
-                  android:imeOptions="actionNext"
-                  android:inputType="textPassword"
-                  android:textAppearance="?android:attr/textAppearanceMedium"
-                  android:textColor="@android:color/black"
-                  />
-
-            </android.support.design.widget.TextInputLayout>
-
-            <android.support.design.widget.TextInputLayout
-                android:id="@+id/pin_hint"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:gravity="center_horizontal"
-                android:hint="@string/optional_lock_code_hint"
-                >
-
-              <com.pyamsoft.pydroid.design.widget.NoLeakTextInputEditText
-                  android:id="@+id/pin_hint_text"
-                  android:layout_width="match_parent"
-                  android:layout_height="match_parent"
-                  android:gravity="center_horizontal"
-                  android:imeOptions="actionGo"
-                  android:inputType="text"
-                  android:textAppearance="?android:attr/textAppearanceSmall"
-                  android:textColor="@android:color/black"
-                  />
-
-            </android.support.design.widget.TextInputLayout>
-
-          </LinearLayout>
-
-        </ScrollView>
-
-      </RelativeLayout>
+          android:layout_height="match_parent"
+          />
 
     </LinearLayout>
